+++ BAZProject/Moduls/Trending/Extensions/TrendingViewController+ErrorPageViewDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  TrendingViewController+ErrorPageViewDelegate.swift
+//  BAZProject
+//
+//  Created by 1058889 on 26/01/23.
+//
+
+extension TrendingViewController: ErrorPageViewDelegate {
+    func reintent() {
+        getData()
+    }
+}
