+++ BAZProject/Moduls/Trending/Extensions/TrendingViewController+HeaderBarViewControllerDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  TrendingViewController+HeaderBarViewControllerDelegate.swift
+//  BAZProject
+//
+//  Created by 1058889 on 26/01/23.
+//
+
+extension TrendingViewController: HeaderBarViewControllerDelegate {
+    func resizeView(heightHeader: Double) {
+        heightNavbarView.constant = heightHeader
+    }
+}
