+++ BAZProject/Moduls/Trending/Extensions/TrendingViewController+UITableExtensions.swift
@@ -1,0 +1,34 @@
+//
+//  TrendingViewController+UITableExtensions.swift
+//  BAZProject
+//
+//  Created by 1058889 on 26/01/23.
+//
+
+import UIKit
+
+extension TrendingViewController: UITableViewDelegate, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return movies.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        if let cell = tableView.dequeueReusableCell(withIdentifier: CellMovie.identifier) as? CellMovie {
+            let movie = movies[indexPath.row]
+            let imageUrl: String = (movie.backdropPath ?? "").getUrlImage(sizeImage: .w500)
+            cell.setData(title: movie.title ?? "", imageUrl: imageUrl)
+            return cell
+        }
+        return UITableViewCell()
+    }
+    
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return UITableView.automaticDimension
+    }
+    
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section:Int) -> String?
+    {
+        return mediaType.getMediaTypeTitle()
+    }
+    
+}
