+++ BAZProject/Moduls/Trending/View/TrendingViewController.swift
@@ -1,12 +1,12 @@
 //
-//  TrendingViewController.swift
+//  TrendingViewController+HeaderBarViewControllerDelegate.swift
 //  BAZProject
 //
 //
 
 import UIKit
 
-class TrendingViewController: UIViewController, TrendingViewProtocol, ErrorPageViewProtocol {
+class TrendingViewController: UIViewController, TrendingViewProtocol {
     
     var movies: [MovieResult] = []
     let mediaType: MediaType = .movie
@@ -50,42 +50,4 @@ class TrendingViewController: UIViewController, TrendingViewProtocol, ErrorPageV
         self.errorUIView.isHidden = showErrorPage
     }
     
-    func reintent() {
-        getData()
-    }
-    
-}
-
-// MARK: - TableView's DataSource
-
-extension TrendingViewController: UITableViewDelegate, UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return movies.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        if let cell = tableView.dequeueReusableCell(withIdentifier: CellMovie.identifier) as? CellMovie {
-            let movie = movies[indexPath.row]
-            let imageUrl: String = (movie.backdropPath ?? "").getUrlImage(sizeImage: .w500)
-            cell.setData(title: movie.title ?? "", imageUrl: imageUrl)
-            return cell
-        }
-        return UITableViewCell()
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return UITableView.automaticDimension
-    }
-    
-    func tableView(_ tableView: UITableView, titleForHeaderInSection section:Int) -> String?
-    {
-        return mediaType.getMediaTypeTitle()
-    }
-    
-}
-
-extension TrendingViewController: HeaderBarViewControllerDelegate {
-    func resizeView(heightHeader: Double) {
-        heightNavbarView.constant = heightHeader
-    }
 }
